requirements.txt
Flask==0.11.1
flask-restplus==0.9.2
Flask-Sockets==0.2.1
gevent==1.1.2
gevent-websocket==0.9.5
numpy==1.11.3
scipy==0.18.1
pandas==0.19.2
matplotlib==1.5.3
Pillow==4.0.0
json-cfg==0.4.1
future==0.16.0
backports.csv==1.0.5
